+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
 ### Added
  - Conferencing Check In/Check Out support
  - ZoomRoom MuteUserOnEntry API
@@ -19,6 +20,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
  - Zoom Call Lock Converter/Response to Call Configuration Converter/Response
  - Zoom Updated Call Record Info Event Converter/Response
  - Added method to IFavorites for getting favorites by dial protocol
+
+### Changed
+ - Renamed old appearances of "Source" to "Participant", and "SourceType" to "CallType"
 
 ## [13.8.1] - 2019-10-08
 ### Changed